requirements.txt
flask
python-docx
num2words
